+++ packages/ai/src/providers/index.ts
@@ -7,6 +7,7 @@ import { MistralAIProvider } from "./mistral";
 import { GoogleAIProvider } from "./google";
 import { AnthropicAIProvider } from "./anthropic";
 import { FireworksAIProvider } from "./fireworks";
+import { AzureOpenAIProvider } from "./azure-openai";
 
 export const chatProvider = new Map<string, ICreateChatCompletion>([
   [OpenAIProvider.name, OpenAIProvider.chat],
@@ -14,6 +15,7 @@ export const chatProvider = new Map<string, ICreateChatCompletion>([
   [GoogleAIProvider.name, GoogleAIProvider.chat],
   [AnthropicAIProvider.name, AnthropicAIProvider.chat],
   [FireworksAIProvider.name, FireworksAIProvider.chat],
+  [AzureOpenAIProvider.name, AzureOpenAIProvider.chat],
 ]);
 
 export const assistantProvider = new Map<string, ICreateAndRunAssistantThread>([